requirements.txt
Django
djangorestframework
psycopg2-binary
pandas
openpyxl
celery
redis
python-dateutil # <--- ADD THIS LINE